Исправил BotService
