+++ src/main/java/skypro/teamwork/telegram_bot_for_shelter/service/BotService.java
@@ -9,6 +9,9 @@ import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageTe
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.buttons.InlineKeyboardButton;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+import skypro.teamwork.telegram_bot_for_shelter.service.button.ButtonCatService;
+import skypro.teamwork.telegram_bot_for_shelter.service.button.ButtonDogService;
+import skypro.teamwork.telegram_bot_for_shelter.service.button.ButtonService;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -48,7 +51,6 @@ public class BotService {
                 ButtonService.callbackQueryAfterCommandStart);
         buttonService.responseOnPressButton(chatId, messageId, msgText, inlineKeyboard);
     }
-
 
 
     /**
@@ -224,9 +226,6 @@ public class BotService {
         );
         buttonService.responseOnPressButton(chatId, messageId, textVaultService.listOfReasonForAdoptingCat, inlineKeyboardDog);
     }
-
-
-
 
 
     /**
@@ -425,413 +424,5 @@ public class BotService {
                 ButtonDogService.callbackQueryAfterCommandGroupHowTakeDog
         );
         buttonService.responseOnPressButton(chatId, messageId, textVaultService.cynologistAdviceDog, inlineKeyboardDog);
-    }
-
-    @Service
-    public static class ButtonCatService {
-        /**
-         * ArrayList c кнопками меню после нажатия кнопки "CHOOSE_A_SHELTER_CAT"
-         */
-        public static final List<String> textButtonsAfterCommandCat = new ArrayList<>(List.of(
-                "О приюте кошек",
-                "Как взять питомца из приюта",
-                "Прислать отчет о питомце",
-                "Позвать волонтёра",
-                "Назад"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопки "CHOOSE_A_SHELTER_CAT"
-         */
-        public static final List<String> callbackQueryAfterCommandCat = new ArrayList<>(List.of(
-                "ABOUT_SHELTER_CAT",
-                "HOW_TAKE_CAT",
-                "SEND_REPORT_CAT",
-                "VOLUNTEER",
-                "START_BUTTON_FOR_EDIT_MESSAGE"
-        ));
-
-
-
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопки "ABOUT_SHELTER_CAT"
-         */
-
-        public static final List<String> textButtonsAfterCommandAboutShelterCat = new ArrayList<>(List.of(
-                "Рассказ о приюте",
-                "Расписание и адрес",
-                "Контактные данные охраны",
-                "Рекомандация по технике безопасности на территории",
-                "Принять контактные данные для связи",
-                "Вызвать волонтёра",
-                "Назад"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопки "ABOUT_SHELTER_CAT"
-         */
-        public static final List<String> callbackQueryAfterCommandAboutShelterCat = new ArrayList<>(List.of(
-                "HISTORY_OF_SHELTER_CAT",
-                "SCHEDULE_AND_ADDRESS_CAT",
-                "CONTACT_SECURITY_CAT",
-                "RECOMMENDATION_LEAFY_CAT",
-                "TAKE_CONTACT_FOR_FEEDBACK",
-                "VOLUNTEER",
-                "CHOOSE_A_SHELTER_CAT"
-        ));
-
-
-
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопки "HOW_TAKE_CAT"
-         */
-        public static final List<String> textButtonsAfterCommandHowTakeCat = new ArrayList<>(List.of(
-                "Правила знакомства с животным",
-                "Список документов для опеки",
-                "Рекоммендации по транспортировке",
-                "Рекомандации по обустройству дома для юного питомца",
-                "Рекомандации по обустройству дома для взрослого питомца",
-                "Рекомандации по обустройству дома для питомца с ограниченными возможностями",
-                "Список причин отказа в опеке",
-                "Принять контактные данные для связи",
-                "Вызвать волонтёра",
-                "Назад"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопки "HOW_TAKE_CAT"
-         */
-        public static final List<String> callbackQueryAfterCommandHowTakeCat = new ArrayList<>(List.of(
-                "DATING_RULES_CAT",
-                "LIST_OF_DOCUMENTS_FOR_ADOPTING_A_CAT",
-                "RECOMMENDATIONS_FOR_TRANSPORTATION_CAT",
-                "RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_LITTLE_CAT",
-                "RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_ADULT_CAT",
-                "RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_DISABLED_CAT",
-                "LIST_OF_REASONS_FOR_ADOPTING_CAT",
-                "TAKE_CONTACT_FOR_FEEDBACK",
-                "VOLUNTEER",
-                "CHOOSE_A_SHELTER_CAT"
-        ));
-
-
-
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопок выдаваемых кнопкой  ABOUT_SHELTER_CAT"
-         * список кнопок
-         * HISTORY_OF_SHELTER_CAT
-         * SCHEDULE_AND_ADDRESS_CAT
-         * CONTACT_SECURITY_CAT
-         * RECOMMENDATION_LEAFY_CAT"
-         */
-        public static final List<String> textButtonsAfterCommandGroupAboutShelterCat = new ArrayList<>(List.of(
-                "Назад",
-                "Вернуться в начало"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопок выдаваемых кнопкой  ABOUT_SHELTER_CAT"
-         * список кнопок
-         * HISTORY_OF_SHELTER_CAT
-         * SCHEDULE_AND_ADDRESS_CAT
-         * CONTACT_SECURITY_CAT
-         * RECOMMENDATION_LEAFY_CAT"
-         */
-        public static final List<String> callbackQueryAfterCommandGroupAboutShelterCat = new ArrayList<>(List.of(
-                "ABOUT_SHELTER_CAT",
-                "START_BUTTON_FOR_EDIT_MESSAGE"
-        ));
-
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопок выдаваемых кнопкой  HOW_TAKE_CAT"
-         * список кнопок
-         * DATING_RULES_CAT
-         * LIST_OF_DOCUMENTS_FOR_ADOPTING_A_CAT
-         * RECOMMENDATIONS_FOR_TRANSPORTATION_CAT
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_LITTLE_CAT
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_ADULT_CAT
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_DISABLED_CAT
-         * LIST_OF_REASONS_FOR_ADOPTING_CAT
-         */
-        public static final List<String> textButtonsAfterCommandGroupHowTakeCat = new ArrayList<>(List.of(
-                "Назад",
-                "Вернуться в начало"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопок выдаваемых кнопкой  HOW_TAKE_CAT"
-         * список кнопок
-         * DATING_RULES_CAT
-         * LIST_OF_DOCUMENTS_FOR_ADOPTING_A_CAT
-         * RECOMMENDATIONS_FOR_TRANSPORTATION_CAT
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_LITTLE_CAT
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_ADULT_CAT
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_DISABLED_CAT
-         * LIST_OF_REASONS_FOR_ADOPTING_CAT
-         */
-        public static final List<String> callbackQueryAfterCommandGroupHowTakeCat = new ArrayList<>(List.of(
-                "HOW_TAKE_CAT",
-                "START_BUTTON_FOR_EDIT_MESSAGE"
-        ));
-    }
-
-    @Service
-    public static class ButtonDogService {
-        /**
-         * ArrayList c кнопками меню после нажатия кнопки "CHOOSE_A_SHELTER_DOG"
-         */
-        public static final List<String> textButtonsAfterCommandDog = new ArrayList<>(List.of(
-                "О собачьем приюте",
-                "Как взять питомца из приюта",
-                "Прислать отчет о питомце",
-                "Позвать волонтёра",
-                "Назад"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопки "CHOOSE_A_SHELTER_DOG"
-         */
-        public static final List<String> callbackQueryAfterCommandDog = new ArrayList<>(List.of(
-                "ABOUT_SHELTER_DOG",
-                "HOW_TAKE_DOG",
-                "SEND_REPORT_DOG",
-                "VOLUNTEER",
-                "START_BUTTON_FOR_EDIT_MESSAGE"
-        ));
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопки "ABOUT_SHELTER_DOG"
-         */
-
-        public static final List<String> textButtonsAfterCommandAboutShelterDog = new ArrayList<>(List.of(
-                "Рассказ о приюте",
-                "Расписание и адрес",
-                "Контактные данные охраны",
-                "Рекомандация по технике безопасности на территории",
-                "Принять контактные данные для связи",
-                "Вызвать волонтёра",
-                "Назад"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопки "ABOUT_SHELTER_DOG"
-         */
-        public static final List<String> callbackQueryAfterCommandAboutShelterDog = new ArrayList<>(List.of(
-                "HISTORY_OF_SHELTER_DOG",
-                "SCHEDULE_AND_ADDRESS_DOG",
-                "CONTACT_SECURITY_DOG",
-                "RECOMMENDATION_LEAFY_DOG",
-                "TAKE_CONTACT_FOR_FEEDBACK",
-                "VOLUNTEER",
-                "CHOOSE_A_SHELTER_DOG"
-        ));
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопки "HOW_TAKE_DOG"
-         */
-        public static final List<String> textButtonsAfterCommandHowTakeDog = new ArrayList<>(List.of(
-                "Правила знакомства с животным",
-                "Список документов для опеки",
-                "Рекоммендации по транспортировке",
-                "Рекомандации по обустройству дома для юного питомца",
-                "Рекомандации по обустройству дома для взрослого питомца",
-                "Рекомандации по обустройству дома для питомца с ограниченными возможностями",
-                "Список причин отказа в опеке",
-                "Список проверенных кинологов",
-                "Советы кинологов по первичному общению с питомцем",
-                "Принять контактные данные для связи",
-                "Вызвать волонтёра",
-                "Назад"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопки "HOW_TAKE_DOG"
-         */
-        public static final List<String> callbackQueryAfterCommandHowTakeDog = new ArrayList<>(List.of(
-                "DATING_RULES_DOG",
-                "LIST_OF_DOCUMENTS_FOR_ADOPTING_DOG",
-                "RECOMMENDATIONS_FOR_TRANSPORTATION_DOG",
-                "RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_LITTLE_DOG",
-                "RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_ADULT_DOG",
-                "RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_DISABLED_DOG",
-                "LIST_OF_REASONS_FOR_ADOPTING_DOG",
-                "LIST_OF_VERIFIED_CYNOLOGISTS_DOG",
-                "CYNOLOGISTS_ADVICE_DOG",
-                "TAKE_CONTACT_FOR_FEEDBACK",
-                "VOLUNTEER",
-                "CHOOSE_A_SHELTER_DOG"
-        ));
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопок выдаваемых кнопкой  ABOUT_SHELTER_DOG"
-         * список кнопок
-         * HISTORY_OF_SHELTER_DOG
-         * SCHEDULE_AND_ADDRESS_DOG
-         * CONTACT_SECURITY_DOG
-         * RECOMMENDATION_LEAFY_DOG"
-         */
-        public static final List<String> textButtonsAfterCommandGroupAboutShelterDog = new ArrayList<>(List.of(
-                "Назад",
-                "Вернуться в начало"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопок выдаваемых кнопкой  ABOUT_SHELTER_DOG
-         * список кнопок
-         * HISTORY_OF_SHELTER_DOG
-         * SCHEDULE_AND_ADDRESS_DOG
-         * CONTACT_SECURITY_DOG
-         * RECOMMENDATION_LEAFY_DOG"
-         */
-        public static final List<String> callbackQueryAfterCommandGroupAboutShelterDog = new ArrayList<>(List.of(
-                "ABOUT_SHELTER_DOG",
-                "START_BUTTON_FOR_EDIT_MESSAGE"
-        ));
-
-
-        /**
-         * ArrayList c кнопками меню после нажатия кнопок выдаваемых кнопкой  HOW_TAKE_DOG
-         * список кнопок
-         * DATING_RULES_DOG
-         * LIST_OF_DOCUMENTS_FOR_ADOPTING_DOG
-         * RECOMMENDATIONS_FOR_TRANSPORTATION_DOG
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_LITTLE_DOG
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_ADULT_DOG
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_DISABLED_DOG
-         * LIST_OF_REASONS_FOR_ADOPTING_DOG
-         * LIST_OF_VERIFIED_CYNOLOGISTS
-         * CYNOLOGISTS_ADVICE
-         */
-        public static final List<String> textButtonsAfterCommandGroupHowTakeDog = new ArrayList<>(List.of(
-                "Назад",
-                "Вернуться в начало"
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после нажатия кнопок выдаваемых кнопкой  HOW_TAKE_DOG
-         * список кнопок
-         * DATING_RULES_DOG
-         * LIST_OF_DOCUMENTS_FOR_ADOPTING_DOG
-         * RECOMMENDATIONS_FOR_TRANSPORTATION_DOG
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_LITTLE_DOG
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_ADULT_DOG
-         * RECOMMENDATIONS_FOR_HOME_DESIGN_FOR_DISABLED_DOG
-         * LIST_OF_REASONS_FOR_ADOPTING_DOG
-         * LIST_OF_REASONS_FOR_ADOPTING_DOG
-         * LIST_OF_VERIFIED_CYNOLOGISTS
-         */
-        public static final List<String> callbackQueryAfterCommandGroupHowTakeDog = new ArrayList<>(List.of(
-                "HOW_TAKE_DOG",
-                "START_BUTTON_FOR_EDIT_MESSAGE"
-        ));
-
-    }
-
-    @Service
-    public static class ButtonService {
-
-        private final Logger logger = LoggerFactory.getLogger(ButtonService.class);
-
-        /**
-         * ArrayList c кнопками меню после команды "start"
-         */
-        public static final List<String> textButtonsAfterCommandStart = new ArrayList<>(List.of(
-                "Приют для кошек ",
-                "Приют для собак "
-        ));
-
-        /**
-         * ArrayList c индификаторами кнопок меню после команды "start"
-         */
-        public static final List<String> callbackQueryAfterCommandStart = new ArrayList<>(List.of(
-                "CHOOSE_A_SHELTER_CAT",
-                "CHOOSE_A_SHELTER_DOG"
-        ));
-
-        private final TelegramBot telegramBot;
-
-        public ButtonService(TelegramBot telegramBot) {
-            this.telegramBot = telegramBot;
-        }
-
-        /**
-         * Метод создает InlineKeyboardMarkup клавиатуру (в каждой строке одна кнопка)
-         *
-         * @param buttonsTexts    текст на кнопках
-         * @param buttonsCallback идентификаторы кнопок
-         * @return инлайн клавитура c кнопками
-         */
-
-        public InlineKeyboardMarkup prepareKeyboard(List<String> buttonsTexts, List<String> buttonsCallback) {
-
-            //создаем клавиатуру
-            InlineKeyboardMarkup markupInline = new InlineKeyboardMarkup();
-
-            //создаем список строк
-            List<List<InlineKeyboardButton>> rowList = new ArrayList<>();
-
-            //в цикле:
-            //создаем кнопки и добавляем их в строку, а строки в лист строк
-            //за один проход цикла создается одна кнопка в каждой строке
-            //количество кнопок определяется входящим параметром buttonsTexts (количеством записей)
-            for (int i = 0; i < buttonsTexts.size(); i++) {
-                List<InlineKeyboardButton> keyboardButtonsRow = new ArrayList<>();
-                InlineKeyboardButton inlineKeyboardButton = new InlineKeyboardButton();
-                inlineKeyboardButton.setText(buttonsTexts.get(i));
-                inlineKeyboardButton.setCallbackData(buttonsCallback.get(i));
-                keyboardButtonsRow.add(inlineKeyboardButton);
-                rowList.add(keyboardButtonsRow);
-            }
-
-            markupInline.setKeyboard(rowList);
-
-            return markupInline;
-        }
-        /**
-         * Метод отправляет пользователю сообщение с клавиатурой под сообщением
-         *
-         * @param chatId         идентификатор чата для отправки сообщения
-         * @param messageText    текст сообщения
-         * @param inlineKeyboard клавиатура под сообщением
-         */
-
-        /** responseStartButton отлаввливает сообщение старт и создает первое сообщение с клавиатурой,
-         * далее кнопки обращаются к методу responseOnPressButton и он заменяет предыдущее сообщение новым с новой клавиатурой
-         */
-        public void responseStartButton(long chatId, String messageText, InlineKeyboardMarkup inlineKeyboard) {
-
-            SendMessage sendMess = new SendMessage(String.valueOf(chatId), messageText);
-            sendMess.setReplyMarkup(inlineKeyboard);
-            try {
-                telegramBot.execute(sendMess);
-            } catch (TelegramApiException e) {
-                logger.error("Произошла ошибка в методе responseStartButton: " + e.getMessage());
-            }
-        }
-
-        public void responseOnPressButton(long chatId, long messageId, String messageText, InlineKeyboardMarkup inlineKeyboard) {
-
-            EditMessageText sendMess = new EditMessageText();
-            sendMess.setChatId(String.valueOf(chatId));
-            sendMess.setText(messageText);
-            sendMess.setMessageId((int)messageId);
-            sendMess.setReplyMarkup(inlineKeyboard);
-            try {
-                telegramBot.execute(sendMess);
-            } catch (TelegramApiException e) {
-                logger.error("Произошла ошибка в методе responseOnPressButton: " + e.getMessage());
-            }
-        }
     }
 }
